Move app.js into /services dir (#75)
